# swagger-static/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user/create_doctor": {
      "post": {
        "operationId": "UserController_create_doctor",
        "summary": "Creates a doctor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDoctorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfull Response",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "msg": "Created Doctor Account"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Access",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/user/get_clinic_info": {
      "get": {
        "operationId": "UserController_get_clinicInfo",
        "summary": "Gets a doctor clinic info",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicInfoDto"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/user/patient_list": {
      "get": {
        "operationId": "UserController_getPatientList",
        "summary": "Gets the list of patients that have had an appointments with the doctor",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientInfo"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/user/edit_doctor_info": {
      "put": {
        "operationId": "UserController_editDoctorInfo",
        "summary": "Edit Doctor info",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditDoctoInfoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfull Response",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "msg": "Doctor changed info"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/user/{user_id}/delete": {
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfull Response",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "msg": "User deleted"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/user/patients": {
      "get": {
        "operationId": "UserController_get_patients",
        "summary": "Gets a list of all the patients in the database",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "test": [
                      {
                        "_id": "string",
                        "email": "string@gmail.com",
                        "firstName": "string",
                        "lastName": "string",
                        "image": "string"
                      },
                      {
                        "_id": "string",
                        "email": "string@gmail.com",
                        "firstName": "string",
                        "lastName": "string",
                        "image": "string"
                      },
                      {
                        "_id": "string",
                        "email": "string@gfr.com",
                        "firstName": "string",
                        "lastName": "string",
                        "image": null
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Access",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/user/patient_details_email/{email}": {
      "get": {
        "operationId": "UserController_get_patient_details_email",
        "summary": "Gets patient details using that patient email",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "_id": "string",
                    "firstName": "string",
                    "lastName": "string",
                    "email": "string@gmail.com",
                    "doctors_id": [
                      "string"
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Access",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/user/patient_details/{id}": {
      "get": {
        "operationId": "UserController_get_patient_details_id",
        "summary": "Gets patient details using that patient _id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "_id": "string",
                    "firstName": "string",
                    "lastName": "string",
                    "email": "string@gmail.com",
                    "doctors_id": [
                      "string"
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Access",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/user/add_note": {
      "post": {
        "operationId": "UserController_add_note",
        "summary": "Adds a note doctor to a patient",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotesDto"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Access",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/user/get_notes/{patient_id}": {
      "get": {
        "operationId": "UserController_get_notes",
        "summary": "Get all the notes for a patient written by a doctor",
        "parameters": [
          {
            "name": "patient_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Access",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/user/update_note/{note_id}": {
      "put": {
        "operationId": "UserController_update_notes",
        "summary": "Updates a note",
        "parameters": [
          {
            "name": "note_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Access",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Theraline endpoints",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ClinicInfoDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "location": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "phone",
          "location"
        ]
      },
      "CreateDoctorDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "birthday": {
            "format": "date-time",
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "clinicInfo": {
            "$ref": "#/components/schemas/ClinicInfoDto"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password",
          "gender",
          "birthday",
          "image",
          "clinicInfo"
        ]
      },
      "Appointment": {
        "type": "object",
        "properties": {}
      },
      "PatientInfo": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "nextAppointment": {
            "$ref": "#/components/schemas/Appointment"
          },
          "lastAppointment": {
            "$ref": "#/components/schemas/Appointment"
          }
        },
        "required": [
          "_id",
          "firstName",
          "lastName",
          "email"
        ]
      },
      "EditDoctoInfoDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "image",
          "phone"
        ]
      },
      "CreateNotesDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "body",
          "user_id"
        ]
      },
...
    }
  }
}